Split configuration types into separate file
